django/requirements.txt
aiohttp==3.7.4.post0
channels==3.0.4
channels-redis==3.3.0
Django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
